+++ src/static/content_scripts/main.js
@@ -14,22 +14,22 @@ function onMessage(message, sender, sendResponse) {
     (_value) => {
       if (message.type == 'PARSE_MUSIC_LIST') {
         console.log('parsing music list ...');
-        sendResponse(Parser.parseMusicList(document.body));
+        sendResponse(Parser.parseMusicList(document.body, message.gameVersion));
         return;
       }
       if (message.type == 'PARSE_MUSIC_DETAIL') {
         console.log('parsing music detail ...');
-        sendResponse(Parser.parseMusicDetail(document.body));
+        sendResponse(Parser.parseMusicDetail(document.body, message.gameVersion));
         return;
       }
       if (message.type == 'PARSE_SCORE_LIST') {
         console.log('parsing score list ...');
-        sendResponse(Parser.parseScoreList(document.body));
+        sendResponse(Parser.parseScoreList(document.body, message.gameVersion));
         return;
       }
       if (message.type == 'PARSE_SCORE_DETAIL') {
         console.log('parsing score detail ...');
-        sendResponse(Parser.parseScoreDetail(document.body));
+        sendResponse(Parser.parseScoreDetail(document.body, message.gameVersion));
         return;
       }
       console.log('received unknown message');